Working on fixing all spring5 check Requests. 1. We overflowed when our unitRadius was in millimeters and we went across the world. We no longer overflow when that occurs. Done by making distances between places longs instead of ints. THIS CHANGES ALL OPTS. 1opt and 2opt have been changed to handle the change from ints to longs. Added Abel's Bio 2. As said, added Abel's Bio to info tab

+++ server/src/main/java/com/tripco/t04/planner/NearestNeighbor.java
@@ -3,15 +3,15 @@ package com.tripco.t04.planner;
 import java.util.ArrayList;
 
 public class NearestNeighbor extends Optimize{
-    NearestNeighbor(ArrayList<Place> places, int[][] latice){
+    NearestNeighbor(ArrayList<Place> places, Long[][] latice){
         super(places, latice);
     }
 
     private int neighbor ( int[] indices, int from){
-        int min = Integer.MAX_VALUE;
+        Long min = Long.MAX_VALUE;
         int j = -1;
         for (int i = from; i < places.size(); i++) {
-            int distance = latice[indices[from - 1]][indices[i]];
+            Long distance = latice[indices[from - 1]][indices[i]];
             if (distance < min) {
                 min = distance;
                 j = i;
@@ -23,11 +23,11 @@ public class NearestNeighbor extends Optimize{
         return j; //returns index of nearest neighbor
     }
 
-    protected int optimizer(int[] candidate){
-        int Total = 0;
+    protected Long optimizer(int[] candidate){
+        Long Total = (long)0;
         for (int i = 1; i < candidate.length; i++) {//for all start points
             int nearestIndex = neighbor(candidate, i);
-            int dist = latice[candidate[i - 1]][candidate[nearestIndex]];
+            Long dist = latice[candidate[i - 1]][candidate[nearestIndex]];
             Total += dist;
             swap(candidate, i, nearestIndex);
         }
